agora v0.20.1: park window accepts fractional seconds; trim park_retire's test window
The suite's remaining floor was park_retire (4.56s): four parking tests each
idled a full 1s waiting out the park window for the re-park block. The window
couldn't go lower because stop.py cast GROUPCHAT_PARK_WINDOW with int()
(int("0.3") raises → silent fallback to the 570s default).

- stop.py: parse the park window as float (whole-second configs still parse —
  float("570")==570.0; junk still falls back to the default; a fractional
  window is now valid). Backward-compatible: no behavior change for any
  existing whole-second value.
- park_retire_test._FAST: PARK_WINDOW 1 → 0.3. The re-park block is emitted at
  the window deadline regardless of poll count, so every assertion holds; the
  four parking tests just stop idling a full second each.

park_retire 4.56 → 1.72s; full suite 4.66 → 3.60s (42/42, stable over 4 runs).
136.55s → 3.60s overall (38x). New floor: hooks_test / push_test (~3.5s).

+++ agora/.groupchat/hooks/stop.py
@@ -60,7 +60,10 @@ def _park_override():
 
 # Poll for this long, then re-park (< the Stop hook timeout of 600s). Env-tunable
 # (and shrunk in tests). Tick is the barrier / @mention detection latency.
-PARK_WINDOW_SECONDS = _envnum("PARK_WINDOW", 570, int)
+# float, not int: whole-second configs still parse (float("570")==570.0), junk still
+# falls back to the default, and a fractional window is now valid (the tests use a
+# sub-second one to avoid idling a full second per parking case).
+PARK_WINDOW_SECONDS = _envnum("PARK_WINDOW", 570, float)
 # Floored to a small positive: a negative env would reach time.sleep() (ValueError,
 # silently defeating the barrier for that agent); 0 would busy-spin the park window.
 POLL_TICK_SECONDS = max(0.05, _envnum("POLL_TICK", 2.0, float))
